fix: Sentiment 엔티티 컬럼 타입 변경

+++ src/main/java/com/example/SentiStock_backend/sentiment/domain/dto/SentimentResponseDto.java
@@ -11,11 +11,11 @@ import lombok.*;
 public class SentimentResponseDto {
 
     private Long id;
-    private Long probPos;
-    private Long probNeg;
-    private Long probNeu;
+    private Double probPos;
+    private Double probNeg;
+    private Double probNeu;
     private String label;
-    private Long score;
+    private Double score;
     private String date;
     private Long newsId;
